pyproject.toml
[tool.poetry]
name = "gochiusa-db-scrape"
version = "0.1.0"
description = ""
authors = ["hoge <hoge@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = {extras = ["cli"], version = "^0.19.0"}
lxml = "^4.8.0"
requests = "^2.27.1"
PyYAML = "^6.0"
pykakasi = "^2.2.1"

[tool.poetry.scripts]
start = "src.main:main"
convert = "src.convert:main"

[tool.poetry.dev-dependencies]
pysen = {version = "0.9.1", extras = ["lint"]}
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 88
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]
